requirements.txt
fastapi
pydantic
pymongo
python-dotenv
uvicorn
email-validator
python-multipart
langchain_google_genai
langchain_mongodb
